revise initialize_sequence function let initialize_sequence can generate large-scale sequence(turn the parameter len from int to long long)

+++ cpu/functions.c
@@ -2,7 +2,7 @@
 #include <limits.h>
 #include <stddef.h>
 
-void initialize_sequence(char **seq, int len){
+void initialize_sequence(char **seq, long long len){
     //鹼基集
     const char baseset[] = "ACGT";
 
@@ -10,7 +10,7 @@ void initialize_sequence(char **seq, int len){
     *seq = (char*)malloc((len + 1) * sizeof(char));
 
     //隨機生成 seq
-    for(int i = 0; i <= len; i++){
+    for(long long i = 0; i <= len; i++){
         (*seq)[i] = baseset[rand() % 4];
     }
